Update to have same hyperparameters in RbRL/PbRL papers
Fix warning in dmc2gym.py by adding int64 rather than int. Updated policy batch, number of envs, how frequent updates occur and n_step for run_ratings.py and run_pref.py. Add the adjusted hyperparameters to run_ppo.py.

+++ dmc2gym.py
@@ -63,7 +63,7 @@ def make(
 def _spec_to_box(spec):
     def extract_min_max(s):
         assert s.dtype == np.float64 or s.dtype == np.float32
-        dim = np.int(np.prod(s.shape))
+        dim = np.int64(np.prod(s.shape))
         if type(s) == specs.Array:
             # Default bound for unbounded action space
             bound = np.inf * np.ones(dim, dtype=np.float32)  # Assuming default bounds of -1 to 1
